update controllers(blog, book, tag, transaction) and transaction model

+++ controller/transaction/transaction.js
@@ -11,8 +11,8 @@ const addTransaction = catchAsync(async (req, res) => {
     data.sender = currentUser?._id;
 
     try {
-        if (!data.sender || data.sender == "" || !data.receiver || data.receiver == "", !data.source || data.source == "", !data) {
-            res.status(STATUS_CODE.ALREADY).json({ message: ERRORS.REQUIRED.FIELD })
+        if (!data.sender || data.sender == "" || !data.receiver || data.receiver == "", !data.source || data.source == "" || !data.orderPrice || data.orderPrice == "" || !data.transactionType || data.transactionType == "" || !data.balance || data.balance == "" || !data.charges || data.charges=="" || !data.orderType || data.orderType == "") {
+            return res.status(STATUS_CODE.ALREADY).json({ message: ERRORS.REQUIRED.FIELD })
         }
         const newData = new transaction(data)
         await newData.save()
@@ -43,9 +43,37 @@ const getAllTransaction = catchAsync(async (req, res) => {
 // This is the Transaction Get One API
 const getTransactionById = catchAsync(async (req, res) => {
     let transactionId = req.params.id
+
     try {
+        const FindOne = await transaction.findById(transactionId)
+        if (!FindOne) {
+            return res.status(STATUS_CODE.BAD_REQUEST).json({ message: ERRORS.INVALID.NOT_FOUND })
+        }
         const result = await transaction.findById(transactionId);
         res.status(STATUS_CODE.OK).json({ message: SUCCESS_MSG.SUCCESS_MESSAGES.SUCCESS, result })
+    } catch (err) {
+        res.status(STATUS_CODE.BAD_REQUEST).json({ message: ERRORS.PROGRAMMING.SOME_ERROR, err })
+    }
+})
+
+// This is update Review Transaction API
+const reviewTransaction = catchAsync(async (req, res) => {
+    const { transactionId, status } = req.body;
+    try {
+        if (!status || status == "") {
+            return res.status(STATUS_CODE.BAD_REQUEST).json({ message: ERRORS.REQUIRED.FIELD })
+        }
+        const FindOne = await transaction.findOne({ _id: transactionId, status: "pending" })
+        if (FindOne) {
+            if (FindOne.status == "approved" || FindOne.status == "rejected") {
+                res.status(STATUS_CODE.BAD_REQUEST).json({ message: SUCCESS_MSG.SUCCESS_MESSAGES.ALREADY })
+            } else {
+                const result = await transaction.findOneAndUpdate({ _id: transactionId }, { $set: { status: status } }, { new: true })
+                return res.status(STATUS_CODE.OK).json({ message: SUCCESS_MSG.SUCCESS_MESSAGES.UPDATE, result })
+            }
+        } else {
+            return res.status(STATUS_CODE.BAD_REQUEST).json({ message: SUCCESS_MSG.SUCCESS_MESSAGES.ALREADY })
+        }
     } catch (err) {
         res.status(STATUS_CODE.BAD_REQUEST).json({ message: ERRORS.PROGRAMMING.SOME_ERROR, err })
     }
@@ -56,6 +84,11 @@ const updateTransactionById = catchAsync(async (req, res) => {
     const updateData = req.body
     const transactionId = req.params.id;
     try {
+        const FindOne = await transaction.findById(transactionId)
+        if (!FindOne) {
+            return res.status(STATUS_CODE.BAD_REQUEST).json({ message: ERRORS.INVALID.NOT_FOUND })
+        }
+        updateData.status = "pending"
         const result = await transaction.findByIdAndUpdate(transactionId, { $set: updateData }, { new: true });
         res.status(STATUS_CODE.OK).json({ message: SUCCESS_MSG.SUCCESS_MESSAGES.UPDATE, result })
     } catch (err) {
@@ -68,6 +101,10 @@ const deleteTransactionById = catchAsync(async (req, res) => {
     const currentUser = req.user
     const transactionId = req.params.id
     try {
+        const FindOne = await transaction.findById(transactionId)
+        if (!FindOne) {
+            return res.status(STATUS_CODE.BAD_REQUEST).json({ message: ERRORS.INVALID.NOT_FOUND })
+        }
         let result;
         if ([ROLES.ADMIN, ROLES.SUPERADMIN].includes(currentUser.role)) {
             result = await transaction.findByIdAndDelete(transactionId);
@@ -85,4 +122,4 @@ const deleteTransactionById = catchAsync(async (req, res) => {
 })
 
 
-module.exports = { addTransaction, getAllTransaction, getTransactionById, updateTransactionById, deleteTransactionById };
+module.exports = { addTransaction, getAllTransaction, getTransactionById, reviewTransaction, updateTransactionById, deleteTransactionById };
